Revert "Transfer all tests to a tests module  (#48)" (#50)
This reverts commit 332dd7e03bc2262a4fcda08446e187713914c7f7.
