+++ CMS/urls.py
@@ -13,16 +13,9 @@ Including another URLconf
     1. Import the include() function: from django.urls import include, path
     2. Add a URL to urlpatterns:  path('blog/', include('blog.urls'))
 """
-from xml.etree.ElementInclude import include
 from django.contrib import admin
-from django.urls import path, include
+from django.urls import path
 
 urlpatterns = [
     path('admin/', admin.site.urls),
-    path('', include('blog.urls')),
-
-    #REST FRAMEWORK URLS
-    path('api-auth/', include('rest_framework.urls')),
-    path('api/post/', include('post.api.urls','blog_api')),
 ]
-
