+++ blog/admin.py
@@ -1,5 +1,3 @@
 from django.contrib import admin
-from .models import Author
-# Register your models here.
 
-admin.site.register([Author])
+# Register your models here.
