--- blog/urls.py
@@ -1,7 +1,0 @@
-from django import views
-from django.urls import path
-from . import views
-
-urlpatterns = [
-    path('', views.home, name='home'),
-]
