+++ blog/views.py
@@ -1,7 +1,3 @@
-from django.http import HttpResponse
-from django.http import HttpResponse
+from django.shortcuts import render
 
 # Create your views here.
-
-def home(request):
-    return HttpResponse('<h1>Hello SpaceyaTech</h1>')
