--- post/admin.py
@@ -1,5 +1,0 @@
-from django.contrib import admin
-from .models import BlogPost
-
-# Register your models here.
-admin.site.register(BlogPost)
