--- post/api/serializers.py
@@ -1,10 +1,0 @@
-
-from rest_framework import serializers
-from post.models import BlogPost
-
-
-class BlogPostSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = BlogPost
-        fields= '__all__'
-
